chore: refactor test tests into languages
