+++ src/lib/monitor.ts
@@ -184,7 +184,9 @@ export async function monitor(
   const policy = await snyk.policy.load(policyLocations, { loose: true });
 
   const target = await getTarget(pkg, meta);
-  const targetFileRelativePath = path.join(path.resolve(root), targetFile || '');
+  const targetFileRelativePath = targetFile
+    ? path.join(path.resolve(root), targetFile)
+    : '';
 
   if (target && target.branch) {
     analytics.add('targetBranch', target.branch);
@@ -292,7 +294,9 @@ export async function monitorGraph(
   const policy = await snyk.policy.load(policyLocations, { loose: true });
 
   const target = await getTarget(pkg, meta);
-  const targetFileRelativePath = path.join(path.resolve(root), targetFile || '');
+  const targetFileRelativePath = targetFile
+    ? path.join(path.resolve(root), targetFile)
+    : '';
 
   if (target && target.branch) {
     analytics.add('targetBranch', target.branch);
